test(release): split runtime artifact checks
